requirements.txt
fastapi
uvicorn
transformers
torch
python-docx
